KeepOld and OldPath options

+++ apply.go
@@ -107,7 +107,6 @@ func Apply(update io.Reader, opts *Options) error {
 	if err != nil {
 		return err
 	}
-	defer fp.Close()
 	_, err = io.Copy(fp, bytes.NewReader(newBytes))
 
 	// if we don't call fp.Close(), windows won't let us move the new executable
@@ -115,7 +114,10 @@ func Apply(update io.Reader, opts *Options) error {
 	fp.Close()
 
 	// this is where we'll move the executable to so that we can swap in the updated replacement
-	oldPath := filepath.Join(updateDir, fmt.Sprintf(".%s.old", filename))
+	oldPath := opts.OldPath
+	if oldPath == "" {
+		oldPath = filepath.Join(updateDir, fmt.Sprintf(".%s.old", filename))
+	}
 
 	// delete any existing old exec file - this is necessary on Windows for two reasons:
 	// 1. after a successful update, Windows can't remove the .old file because the process is still running
@@ -148,11 +150,13 @@ func Apply(update io.Reader, opts *Options) error {
 	}
 
 	// move successful, remove the old binary
-	errRemove := os.Remove(oldPath)
+	if !opts.KeepOld {
+		errRemove := os.Remove(oldPath)
 
-	// windows has trouble with removing old binaries, so hide it instead
-	if errRemove != nil {
-		_ = hideFile(oldPath)
+		// windows has trouble with removing old binaries, so hide it instead
+		if errRemove != nil {
+			_ = hideFile(oldPath)
+		}
 	}
 
 	return nil
@@ -205,6 +209,13 @@ type Options struct {
 	// If nil, treat the update as a complete replacement for the contents of the file at TargetPath.
 	// If non-nil, treat the update contents as a patch and use this object to apply the patch.
 	Patcher Patcher
+
+	// If true, keep the old executable
+	// If false, remove the old executable after update (default)
+	KeepOld bool
+
+	//OldPath defines the path where the old executable is stored after update
+	OldPath string
 }
 
 // CheckPermissions determines whether the process has the correct permissions to
